# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EcoDrop API",
    "version": "1.0.0",
    "description": "Nearest recycling bin lookup API powered by PostGIS."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
      "ReportRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "description": "Clicked map latitude; sent as a decimal text field in FormData."
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "description": "Clicked map longitude; sent as a decimal text field in FormData."
          },
          "reporterName": {
            "type": "string",
            "maxLength": 200,
            "description": "Optional reporter name; empty text is treated as absent."
          },
          "incidentTime": {
            "type": "string",
            "format": "date-time",
            "maxLength": 64,
            "description": "Optional RFC 3339 timestamp with timezone, e.g. 2026-09-15T09:00:00+03:00. Empty text is treated as absent."
          },
          "message": {
            "type": "string",
            "maxLength": 10000,
            "description": "Optional free text; empty text is treated as absent."
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Optional single JPEG, PNG, or WebP image, at most 5 MiB (5,242,880 bytes)."
          }
        }
      },
      "ReportSubmitted": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string",
            "enum": [
              "Report submitted successfully"
            ]
          }
        }
      },
      "CreateBinRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "latitude",
          "longitude"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "pattern": "\\S",
            "description": "Trimmed before storage."
          },
          "address": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "pattern": "\\S",
            "default": null,
            "description": "Trimmed before storage; omit or use null if unknown."
          },
          "type": {
            "type": "string",
            "nullable": true,
            "enum": [
              "general_waste",
              "packaging",
              "glass",
              "paper",
              "textile",
              "electronics",
              "cardboard",
              "bulky_waste",
              "bottle_recycling_machine",
              "yard_waste",
              null
            ],
            "description": "Stable map icon key; null means unknown (use a generic icon).\n- general_waste: Green Bin: general / mixed municipal waste.\n- packaging: Orange Packaging Bin: plastic, metal, and beverage carton packaging.\n- glass: Purple Glass Bin: glass bottles and food jars.\n- paper: Blue Paper Bin: newspapers, magazines, and office paper.\n- textile: Textile Bin: used clothing, shoes, and fabrics.\n- electronics: Electronics / E-Waste Bin: small electronic appliances and devices.\n- cardboard: Cardboard Cage / Bin: large flattened cardboard boxes.\n- bulky_waste: Bulky Waste Container: large items like old furniture and mattresses.\n- bottle_recycling_machine: Bottle Recycling Machine: reverse vending machine (RVM) for deposit bottles.\n- yard_waste: Yard Waste / Pruning: tree branches, leaves, and garden trimmings.",
            "default": null
          },
          "latitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          }
        }
      },
      "Bin": {
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "type",
          "latitude",
          "longitude"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "description": "Display address for the map popup; null if unknown.",
            "example": "Central Park, New York, NY"
          },
          "type": {
            "type": "string",
            "nullable": true,
            "enum": [
              "general_waste",
              "packaging",
              "glass",
              "paper",
              "textile",
              "electronics",
              "cardboard",
              "bulky_waste",
              "bottle_recycling_machine",
              "yard_waste",
              null
            ],
            "description": "Stable map icon key; null means unknown (use a generic icon).\n- general_waste: Green Bin: general / mixed municipal waste.\n- packaging: Orange Packaging Bin: plastic, metal, and beverage carton packaging.\n- glass: Purple Glass Bin: glass bottles and food jars.\n- paper: Blue Paper Bin: newspapers, magazines, and office paper.\n- textile: Textile Bin: used clothing, shoes, and fabrics.\n- electronics: Electronics / E-Waste Bin: small electronic appliances and devices.\n- cardboard: Cardboard Cage / Bin: large flattened cardboard boxes.\n- bulky_waste: Bulky Waste Container: large items like old furniture and mattresses.\n- bottle_recycling_machine: Bottle Recycling Machine: reverse vending machine (RVM) for deposit bottles.\n- yard_waste: Yard Waste / Pruning: tree branches, leaves, and garden trimmings.",
            "example": "paper"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "distanceMeters": {
            "type": "number"
          }
        }
      },
      "NearestBinResponse": {
        "type": "object",
        "required": [
          "nearestBin",
          "latitude",
          "longitude",
          "distanceMeters"
        ],
        "properties": {
          "nearestBin": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Bin"
              },
              {
                "type": "object",
                "required": [
                  "distanceMeters"
                ]
              }
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "distanceMeters": {
            "type": "number"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Identity": {
        "type": "object",
        "required": [
          "name",
          "version",
          "environment"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "ecodrop-api"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "environment": {
            "type": "string",
            "example": "development"
          }
        }
      },
      "Health": {
        "type": "object",
        "required": [
          "status",
          "service",
          "version",
          "uptimeSeconds"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ]
          },
          "service": {
            "type": "string",
            "example": "ecodrop-api"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "uptimeSeconds": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "Readiness": {
        "type": "object",
        "required": [
          "status",
          "service",
          "version",
          "database"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "not_ready"
            ]
          },
          "service": {
            "type": "string",
            "example": "ecodrop-api"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "database": {
            "type": "object",
            "required": [
              "status",
              "latencyMs",
              "currentTime",
              "serverVersion"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "connected",
                  "disconnected"
                ]
              },
              "latencyMs": {
                "type": "number",
                "description": "Database probe duration in milliseconds."
              },
              "currentTime": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "serverVersion": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "error": {
            "type": "string",
            "description": "Present when the readiness probe fails."
          }
        }
      }
    }
  },
  "paths": {
    "/whoami": {
      "get": {
        "summary": "Application identity",
        "responses": {
          "200": {
            "description": "Service name, version, and environment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Process liveness (does not query the database)",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Database readiness, query timing, and PostgreSQL version",
        "responses": {
          "200": {
            "description": "Database is reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          },
          "503": {
            "description": "Database readiness check failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            }
          }
        }
      }
    },
    "/api/bins": {
      "get": {
        "summary": "List all recycling bins",
        "responses": {
          "200": {
            "description": "A collection of bins",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bin"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database query or bin data validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create and persist a recycling bin",
        "description": "Public write endpoint. Each request creates a new bin; duplicate names are allowed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBinRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created bin with its database-generated ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bin"
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or bin fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds the 100 KB limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported content type, charset, or encoding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create a bin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bins/nearest": {
      "get": {
        "summary": "Find the nearest recycling bin",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "description": "Latitude in decimal degrees"
          },
          {
            "in": "query",
            "name": "lng",
            "required": true,
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "description": "Longitude in decimal degrees"
          }
        ],
        "responses": {
          "200": {
            "description": "The nearest recycling bin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NearestBinResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid latitude or longitude",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No recycling bins were found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database query or bin data validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports": {
      "post": {
        "summary": "Submit a map incident report for background email delivery",
        "description": "Requires only the clicked location. Responds after upload/validation without waiting for email delivery. Best effort only; a 200 response does not guarantee delivery.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              },
              "encoding": {
                "image": {
                  "contentType": "image/jpeg, image/png, image/webp"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report accepted for background dispatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSubmitted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields, duplicate fields, image signature, or multipart data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Image exceeds 5 MiB or a text field exceeds its upload limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Expected multipart/form-data with an optional JPEG, PNG, or WebP image",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Report email configuration is unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}